+++ postmarker/django/backend.py
@@ -1,0 +1,29 @@
+# coding: utf-8
+from django.conf import settings
+from django.core.exceptions import ImproperlyConfigured
+from django.core.mail.backends.base import BaseEmailBackend
+
+from ..core import ServerClient
+
+
+DEFAULT_CONFIG = {
+    'TOKEN': None,
+}
+
+
+class EmailBackend(BaseEmailBackend):
+    """
+    A wrapper that manages sending emails via Postmark API.
+    """
+
+    def __init__(self, token=None, fail_silently=False, **kwargs):
+        super(EmailBackend, self).__init__(fail_silently=fail_silently)
+        postmark_config = getattr(settings, 'POSTMARK', DEFAULT_CONFIG)
+        self.token = token or postmark_config.get('TOKEN')
+        if self.token is None:
+            raise ImproperlyConfigured('You should specify TOKEN to use Postmark email backend')
+
+    def send_messages(self, email_messages):
+        server_client = ServerClient(token=self.token)
+        prepared_messages = [message.message() for message in email_messages]
+        return len(server_client.emails.send_batch(*prepared_messages))