+++ tests/django/settings.py
@@ -1,0 +1,8 @@
+# coding: utf-8
+
+SECRET_KEY = 'foo'
+
+EMAIL_BACKEND = 'postmarker.django.backend.EmailBackend'
+POSTMARK = {
+    'TOKEN': '<YOUR POSTMARK SERVER TOKEN>'
+}